+++ AR PlayerPic/Assets/ChromaKeyShader.shader.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: a6f1f9b4ae69a4b4fbc8e02770ff02e1
+ShaderImporter:
+  externalObjects: {}
+  defaultTextures: []
+  nonModifiableTextures: []
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
