+++ AR PlayerPic/Assets/Editor/Vuforia/ImageTargetTextures/Database/Messi_scaled.jpg.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: da5c63929bd74ba8a2b01c02b9af3bc0
+guid: 24baa74cc397425d90ed9b0bcc838af5
 TextureImporter:
   fileIDToRecycleName: {}
   externalObjects: {}
