+++ AR PlayerPic/Assets/Editor/Vuforia/ImageTargetTextures/Database/malen_scaled.jpg.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: a8fd54ec46ea477cb04658ef275df3c9
+guid: 9ef3154bf1fd4281af3d54b2a4798b51
 TextureImporter:
   fileIDToRecycleName: {}
   externalObjects: {}
