+++ AR PlayerPic/Assets/Kip.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8ea910684ec228949acd702d591d18a8
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
