+++ AR PlayerPic/Assets/KipShader.shader.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: b7ea1a7169a744d4392272d6881a89dc
+ShaderImporter:
+  externalObjects: {}
+  defaultTextures: []
+  nonModifiableTextures: []
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
