+++ AR PlayerPic/Assets/Scripts/PlayerWalk.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 938f127c7d4c6c341a5dd16138e15a58
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
