+++ AR PlayerPic/Assets/Scripts/restartScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f8598db9c543e2d4fbe80b3f28a36b8b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
