+++ AR PlayerPic/Assets/StreamingAssets/Vuforia/Database.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <QCARConfig xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="qcar_config.xsd">
   <Tracking>
+    <ImageTarget name="ronaldo" size="1.000000 0.562500" />
     <ImageTarget name="Messi" size="1.000000 1.501502" />
     <ImageTarget name="malen" size="700.000000 467.002869" />
   </Tracking>
