+++ AR PlayerPic/Assets/StreamingAssets/Vuforia/Database.xml.meta
@@ -1,2 +1,2 @@
 fileFormatVersion: 2
-guid: 7d8be77352fd4178a28c744dceebd915
+guid: ce2576afd758451398550e7bba9479d5
